+++ runtime-attach/runtime-attach-core/src/main/java/io/opentelemetry/contrib/attach/core/CoreRuntimeAttach.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.attach;
+package io.opentelemetry.contrib.attach.core;
 
 import java.io.File;
 import java.lang.management.ManagementFactory;
